+++ tests/test_spi_pio/main.cpp
@@ -98,15 +98,6 @@ static void _error_blink(int count)
     }
 }
 
-// uncomment below with edit if non-default GPIO attributes are needed
-/*
-extern "C" {
-void pico_fatfs_init_spi(void)
-{
-}
-}
-*/
-
 int main()
 {
     FATFS fs;
@@ -161,17 +152,25 @@ int main()
     printf("=====================\n");
 
     // modify below if customized configuration is needed
+    //   Pin assignments for Pimoroni Pico DV demo base board
     pico_fatfs_spi_config_t config = {
-        spi0,
+        spi0,  // if unmatched SPI pin assignments with spi0/spi1 or explicitly designated as NULL, SPI PIO will be configured 
         CLK_SLOW_DEFAULT,
-        CLK_FAST_DEFAULT,
-        PIN_SPI0_MISO_DEFAULT,
-        PIN_SPI0_CS_DEFAULT,
-        PIN_SPI0_SCK_DEFAULT,
-        PIN_SPI0_MOSI_DEFAULT,
-        true  // use internal pullup
+        CLK_FAST_DEFAULT_PIO,
+        19,    // MISO (SPIx_RX)
+        22,    // CS
+         5,    // SCK  (SPIx_SCK)
+        18,    // MOSI (SPIx_TX)
+        true   // use internal pullup
     };
-    pico_fatfs_set_config(&config);
+    bool spi_configured = pico_fatfs_set_config(&config);
+    if (spi_configured) {
+        printf("SPI configured\n");
+    } else {
+        // modify if customized configuration for SPI PIO is needed
+        pico_fatfs_config_spi_pio(SPI_PIO_DEFAULT_PIO, SPI_PIO_DEFAULT_SM);
+        printf("SPI PIO configured\n");
+    }
 
     for (int i = 0; i < 5; i++) {
         fr = f_mount(&fs, "", 1);